requirements.txt
streamlit
nltk
chardet
PyMuPDF
google-generativeai
sumy
python-dotenv